refact: separate parseXML
